Add files via upload
This is the first working copy. This software controls two servos via the keyboard or via Bluetooth. It runs on a Raspberry Pi using a Pi Hat. The onboard servos had too much jitter. This required the use of the i2c bus. The arrow keys and the w,a,s & z keys control the servo.